Deduplicate native-auth controllers and wire HAL isError propagation
Share token-request handling between the token controller and the V2 flow
controller via a MSALNativeAuthTokenRequestHandling protocol with a default
implementation, keeping performTokenRequest off the base controller. Extract
the request interceptor bridge into a single shared type used by both the V1
and V2 request configurators. Route V2 response errors through isError so the
server error is surfaced up the chain in all validator paths, and drop unused
HALResource link accessors.
